Move overlap handling
